Report missing symbols at compile time, even internal symbols
For some reason we were not reporting missing symbols that started
with `emscripten_` which meant that such missing symbols were not
reported until runtime.

The changes to gl.c were machine generated based on error received
when building with EGACY_GL_EMULATION.  These represent legacy GL
symbols for which we do not yet have any emulation.